requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
cffi==1.15.1
Django==4.2.4
djangorestframework==3.14.0
Pillow==10.0.0
pycparser==2.21
PyMySQL==1.1.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
